# 8279216: Investigate implementation of premultiplied alpha in the Little-CMS 2.13