--- a/src/java.desktop/share/classes/sun/java2d/cmm/lcms/LCMSTransform.java
+++ b/src/java.desktop/share/classes/sun/java2d/cmm/lcms/LCMSTransform.java
@@ -130,12 +130,12 @@ final class LCMSTransform implements ColorTransform {
      * Returns {@code true} if lcms may support this format directly.
      */
     private static boolean isLCMSSupport(BufferedImage src, BufferedImage dst) {
-        if (!dst.getColorModel().hasAlpha()) {
-            return true;
-        }
+        boolean dstAlpha = dst.getColorModel().hasAlpha();
+        boolean srcAlpha = src.getColorModel().hasAlpha();
+        boolean srcPre = srcAlpha && src.getColorModel().isAlphaPremultiplied();
+        // lcms does not convert pre-alpha for transparent src if dst is opaque
         // lcms does not set correct alpha for transparent dst if src is opaque
-        // is it feature or bug?
-        return dst.getColorModel().hasAlpha() == src.getColorModel().hasAlpha();
+        return !dstAlpha && !srcPre || dstAlpha == srcAlpha;
     }
 
     public void colorConvert(BufferedImage src, BufferedImage dst) {