… test/jdk/sun/java2d/cmm/ColorConvertOp/SkipSampleModel.java