Support of premultiplied alpha is added to the "accelerated" code path of the CMM. It is implemented on top of the new feature added to the littlecms library in 2.13 and 2.15.

The next formats are now "supported": TYPE_INT_ARGB_PRE, TYPE_4BYTE_ABGR_PRE, and any custom images which use the ComponentColorModel+8-bit-precision like RGBApre or ApreBGR.

After this patch, we will fully support all our standard types(8-bit precision), and mostly any combinations of blits between them, having two exceptions:
 * lcms does not convert pre-alpha for transparent src if dst is opaque
 * lcms does not set correct alpha(=1.0) for transparent dst if src is opaque

Both of them are "features", so we probably need to implement a workaround someday, but for now we will use our generic/slow code.

Performance results for the ColorConvertOp.filter() for images with premultiplied alpha.

| Test | Base(avgt in us/op)  | Fix(avgt in us/op) | Improvement % |
| ------------- | ------------- | ------------- |------------- |
| 32 Threads, from=sRGB:size=10:to=CIEXYZ  | 97| 12| 708% |
| 32 Threads, from=sRGB:size=100:to=CIEXYZ  | 8 013  | 617  | 1199% |
| 32 Threads, from=sRGB:size=1000:to=CIEXYZ  | 651 838  | 104 969|  521% |
| 1 Thread, from=sRGB:size=10:to=CIEXYZ  | 21| 5|  320% |
| 1 Thread, from=sRGB:size=100:to=CIEXYZ  | 677  | 300 |  126% |
| 1 Thread, from=sRGB:size=1000:to=CIEXYZ  | 60 095  | 30 536 | 97% |

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8279216](https://bugs.openjdk.org/browse/JDK-8279216): Investigate implementation of premultiplied alpha in the Little-CMS 2.13


### Reviewers
 * [Phil Race](https://openjdk.org/census#prr) (@prrace - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/13095/head:pull/13095` \
`$ git checkout pull/13095`

Update a local copy of the PR: \
`$ git checkout pull/13095` \
`$ git pull https://git.openjdk.org/jdk.git pull/13095/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 13095`

View PR using the GUI difftool: \
`$ git pr show -t 13095`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/13095.diff">https://git.openjdk.org/jdk/pull/13095.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/13095#issuecomment-1478632253)