@prrace 2023-04-12
I think the test failures are somewhat spurious,
but the easiest thing to do here is for you to change one line and we'll be OK

< int firstBand = image.getSampleModel().getNumBands() - 1;
> int firstBand = byteRaster.getSampleModel().getNumBands() - 1;