@mrserb 2023-04-13
> ```
> < int firstBand = image.getSampleModel().getNumBands() - 1;
> > int firstBand = byteRaster.getSampleModel().getNumBands() - 1;
> ```
Done. The new test is added to prove that this issue could be reproduced before the current patch.